공연 추천 서버연결 버그 해결

+++ app/src/main/java/com/example/whashow/data/RecommandPerformanceList.kt
@@ -41,7 +41,7 @@ data class PerformancesByStandardList(
     @SerialName("price")
     val price: String,
     @SerialName("ratingAverage")
-    val ratingAverage: Int,
+    val ratingAverage: Float,
     @SerialName("startDate")
     val startDate: String,
     @SerialName("title")
